fix: implement changelogs reordering
